+++ pages/pagesampleddata.h
@@ -59,6 +59,7 @@ private slots:
     void on_filterLogScaleBox_toggled(bool checked);
     void on_plotModeBox_currentIndexChanged(int index);
     void on_saveDataButton_clicked();
+    void on_loadDataButton_clicked();
 
 private:
     Ui::PageSampledData *ui;