+++ res/CHANGELOG.md
@@ -15,6 +15,7 @@
 * Added filter-box to example and recent lists in script editors.
 * Limit output size in vesc and lbm terminal to prevent problems when printing too much.
 * Use all 3 measured currents in sampled data instead of calculating one.
+* Added support for loading sampled data CSV files.
 
 ---